Role based access

+++ app/Http/Controllers/API/BlogPostController.php
@@ -89,9 +89,11 @@ class BlogPostController extends Controller
 
         if (Auth::user()->role == 'admin') {
             $data['except'] = 'published';
-            $data['published_at'] = date('Y-m-d H:i:s');
         }
 
+        if (Auth::user()->role == 'admin' || Auth::user()->role == 'author') {
+            $data['published_at'] = date('Y-m-d H:i:s');
+        }
         BlogPost::create($data);
 
         return response()->json([
@@ -138,12 +140,12 @@ class BlogPostController extends Controller
         // check if the user is the same of the one logged in
         $loggedUser = Auth::user();
 
-        if ($loggedUser->id != $request->user_id) {
-            return response()->json([
-                'status' => 'fail',
-                'message' => 'Unauthorized access.'
-            ], 400);
-        }
+        // if ($loggedUser->id != $request->user_id) {
+        //     return response()->json([
+        //         'status' => 'fail',
+        //         'message' => 'Unauthorized access.'
+        //     ], 400);
+        // }
 
         // check if the category id is valid
         $category = BlogCategory::find($request->category_id);
@@ -155,20 +157,26 @@ class BlogPostController extends Controller
             ], 404);
         }
 
-        $post->title = $request->title;
-        $post->slug = Str::slug($request->title);
-        $post->category_id = $request->category_id;
-        $post->user_id = $request->user_id;
-        $post->content = $request->content;
-        // $post->excerpt = $request->excerpt;
-        $post->status = $request->status;
-        $post->save();
+        // check additional conditional to restrict authorized edit
+        if ($loggedUser->id === $post->user_id || Auth::user()->role == 'admin') {
+            $post->title = $request->title;
+            $post->slug = Str::slug($request->title);
+            $post->category_id = $request->category_id;
+            $post->user_id = $request->user_id;
+            $post->content = $request->content;
+            // $post->excerpt = $request->excerpt;
+            $post->status = $request->status;
+            $post->save();
 
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Blog posts updated successfully.'
+            ], 201);
+        }
         return response()->json([
-            'status' => 'success',
-            'message' => 'Blog posts updated successfully.'
-        ], 201);
-
+            'status' => 'fail',
+            'message' => 'You are not allowed to perform this task.'
+        ], 403);
     }
 
     public function blogPostImage(Request $request, int $id)
@@ -196,34 +204,41 @@ class BlogPostController extends Controller
 
         $loggedUser = Auth::user();
 
-        if ($loggedUser->id != $request->user_id) {
-            return response()->json([
-                'status' => 'fail',
-                'message' => 'Unauthorized access.'
-            ], 400);
-        }
+        // if ($loggedUser->id != $request->user_id) {
+        //     return response()->json([
+        //         'status' => 'fail',
+        //         'message' => 'Unauthorized access.'
+        //     ], 400);
+        // }
 
         $imagePath = null;
-        if ($request->hasFile('thumbnail') && $request->file('thumbnail')->isValid()) {
-            $file = $request->file('thumbnail');
 
-            // generate an unique file name
-            $fileName = time().'_'.$file->getClientOriginalName();
+        if ($loggedUser->id === $post->user_id || Auth::user()->role == 'admin') {
+            if ($request->hasFile('thumbnail') && $request->file('thumbnail')->isValid()) {
+                $file = $request->file('thumbnail');
 
-            // move file into storage
-            $file->move(public_path('storage/posts'), $fileName);
+                // generate an unique file name
+                $fileName = time().'_'.$file->getClientOriginalName();
 
-            // image path into database
-            $imagePath = 'storage/posts/'.$fileName;
+                // move file into storage
+                $file->move(public_path('storage/posts'), $fileName);
+
+                // image path into database
+                $imagePath = 'storage/posts/'.$fileName;
+            }
+
+            $post->thumbnail = $imagePath ?? $post->thumbnail;
+            $post->save();
+
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Image updated successfully.'
+            ], 201);
         }
-
-        $post->thumbnail = $imagePath ?? $post->thumbnail;
-        $post->save();
-
         return response()->json([
-            'status' => 'success',
-            'message' => 'Image updated successfully.'
-        ], 201);
+            'status' => 'fail',
+            'message' => 'You are not allowed to perform this task.'
+        ], 403);
     }
 
     /**
@@ -240,11 +255,21 @@ class BlogPostController extends Controller
             ], 404);
         }
 
-        $post->delete();
+        // check if the user is the same of the one logged in
+        $loggedUser = Auth::user();
+
+        if ($loggedUser->id === $post->user_id || Auth::user()->role == 'admin') {
+            $post->delete();
+
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Post deleted successfully.'
+            ], 201);
+        }
 
         return response()->json([
-            'status' => 'success',
-            'message' => 'Post deleted successfully.'
-        ], 201);
+            'status' => 'fail',
+            'message' => 'You are not allowed to perform this task.'
+        ], 403);
     }
 }
